+++ common/src/main/java/net/mehvahdjukaar/supplementaries/configs/CommonConfigs.java
@@ -315,6 +315,8 @@ public class CommonConfigs {
                     .define("interaction_mode", BlackboardBlock.UseMode.BOTH);
             builder.pop();
 
+
+
             builder.push("gravel_bricks");
             GRAVEL_BRICKS_ENABLED = feature(builder);
             builder.pop();
@@ -459,6 +461,7 @@ public class CommonConfigs {
             DOORMAT_ENABLED = feature(builder, ModConstants.DOORMAT_NAME);
             FLINT_BLOCK_ENABLED = feature(builder, ModConstants.FLINT_BLOCK_NAME);
             CANDLE_HOLDER_ENABLED = feature(builder, ModConstants.CANDLE_HOLDER_NAME);
+            FIRE_PIT_ENABLED = feature(builder, ModConstants.FIRE_PIT_NAME);
             //   SPEEDOMETER_ENABLED = feature(builder, ModConstants.SPEEDOMETER_NAME,ModConstants.SPEEDOMETER_NAME, false);
 
             builder.pop();
@@ -564,6 +567,7 @@ public class CommonConfigs {
         public static final Supplier<Boolean> LAPIS_BRICKS_ENABLED;
 
         public static final Supplier<Boolean> CANDLE_HOLDER_ENABLED;
+        public static final Supplier<Boolean> FIRE_PIT_ENABLED;
 
         public static final Supplier<Boolean> HAT_STAND_ENABLED;
         public static final Supplier<Boolean> HAT_STAND_UNRESTRICTED;