+++ common/src/main/java/net/mehvahdjukaar/supplementaries/reg/ModCreativeTabs.java
@@ -108,6 +108,10 @@ public class ModCreativeTabs {
                 ModConstants.SCONCE_NAME,
                 sconces.toArray(Supplier[]::new));
 
+        before(e, Items.CAMPFIRE, CreativeModeTabs.FUNCTIONAL_BLOCKS,
+                ModConstants.FIRE_PIT_NAME,
+                ModRegistry.FIRE_PIT);
+
         before(e, Items.CHAIN, CreativeModeTabs.FUNCTIONAL_BLOCKS,
                 ModConstants.ROPE_NAME,
                 ModRegistry.ROPE);