+++ fabric/src/main/java/net/mehvahdjukaar/supplementaries/mixins/fabric/BushBlockMixin.java
@@ -1,22 +1,19 @@
 package net.mehvahdjukaar.supplementaries.mixins.fabric;
 
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import net.mehvahdjukaar.supplementaries.common.block.blocks.PlanterBlock;
 import net.minecraft.core.BlockPos;
-import net.minecraft.world.item.BowItem;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.BushBlock;
-import net.minecraft.world.level.block.CropBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(BushBlock.class)
 public abstract class BushBlockMixin {
 
-    @Inject(method = "mayPlaceOn", at = @At("HEAD"), cancellable = true)
-    public void mayPlaceOn(BlockState state, BlockGetter level, BlockPos pos, CallbackInfoReturnable<Boolean> cir) {
-        if (state.getBlock() instanceof PlanterBlock) cir.setReturnValue(true);
+    @ModifyReturnValue(method = "mayPlaceOn", at = @At("RETURN"))
+    public boolean mayPlaceOn(boolean original, BlockState state, BlockGetter level, BlockPos pos) {
+        return !original && state.getBlock() instanceof PlanterBlock;
     }
 }